mestoSporta/requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
black==22.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
dadata==21.10.1
Django==4.1.1
dnspython==2.2.1
email-validator==1.3.0
envparse==0.2.0
frozenlist==1.3.1
gino==1.0.1
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
loguru==0.6.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
